+++ Sources/SwiftExtras/Enums/Device.swift
@@ -44,11 +44,55 @@ public enum Device {
     /// Are we running on Carplay?
     var isCarplay: Bool {
 #if canImport(UIKit) && os(iOS)
-        return UIScreen.screens.filter {
-            $0.traitCollection.userInterfaceIdiom == .carPlay
-        }.count >= 1
+        if #available(iOS 16, *) {
+            return UIApplication.shared.connectedScenes.filter {
+                ($0 as? UIWindowScene)?.traitCollection.userInterfaceIdiom == .carPlay
+            }.count >= 1
+        } else {
+            return UIScreen.screens.filter {
+                $0.traitCollection.userInterfaceIdiom == .carPlay
+            }.count >= 1
+        }
 #else
-        return false
+    return false
 #endif
-    }
+}
+
+#if canImport(UIKit)
+/// The device name (e.g., "John’s iPhone")
+static var deviceName: String {
+    UIDevice.current.name
+}
+
+/// The system name (e.g., "iOS")
+static var deviceSystemName: String { UIDevice.current.systemName }
+
+/// The unique vendor identifier (IDFV)
+static var deviceIdentifierForVendor: String { UIDevice.current.identifierForVendor?.uuidString ?? "No ID" }
+
+/// Detect if the device is an iPad
+static var isiPad: Bool { UIDevice.current.userInterfaceIdiom == .pad }
+
+/// Detect if the device is an iPhone
+static var isiPhone: Bool { UIDevice.current.userInterfaceIdiom == .phone }
+
+#if !os(tvOS)
+/// The current battery level (0.0 to 1.0)
+static var deviceBatteryLevel: Float { UIDevice.current.batteryLevel }
+
+/// The current battery state (charging, full, unplugged)
+static var deviceBatteryState: UIDevice.BatteryState { UIDevice.current.batteryState }
+
+#if !os(visionOS)
+/// The device’s orientation (portrait, landscape, etc.)
+static var deviceOrientation: UIDeviceOrientation { UIDevice.current.orientation }
+
+/// Detect if the device is in portrait mode
+static var isPortraitMode: Bool { UIDevice.current.orientation.isPortrait }
+
+/// Detect if the device is in landscape mode
+static var isLandscapeMode: Bool { UIDevice.current.orientation.isLandscape }
+#endif
+#endif
+#endif
 }