+++ Sources/SwiftExtras/Extensions/Color+components.swift
@@ -67,6 +67,18 @@ extension Color {
         )
     }
 
+    /// Get HEX string
+    /// - Returns a string representation of the color in HEXi format.
+    var hex: String {
+        String(
+            format: "#%02x%02x%02x%02x",
+            Int(components.red * 255),
+            Int(components.green * 255),
+            Int(components.blue * 255),
+            Int(components.opacity * 255)
+        )
+    }
+
     /// Get RGB string
     /// - Returns a string representation of the color in RGB format.
     public func rgbString() -> String {