+++ Sources/SwiftExtras/Extensions/View/View+sizeSaver.swift
@@ -20,7 +20,7 @@ struct SaveSizeModifier: ViewModifier {
         content
             .background(
                 GeometryReader { proxy in
-                    if #available(iOS 17.0, macOS 14.0, *) {
+                    if #available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *) {
                         Color.clear
                             .onAppear {
                                 size = proxy.size